// src/components/arcade/bernard-sprite-atlas.generated.ts
export const BERNARD_ATLAS: SpriteAtlas = {
  "url": "/assets/arcade/fighter-bernard-v3.png",
  "poses": {
    "idle": [
      {
        "x": 8,
        "y": 8,
        "w": 160,
        "h": 190,
        "ax": 80,
        "ay": 190
      },
      {
        "x": 259,
        "y": 8,
        "w": 156,
        "h": 190,
        "ax": 78,
        "ay": 190
      },
      {
        "x": 510,
        "y": 8,
        "w": 157,
        "h": 190,
        "ax": 78,
        "ay": 190
      },
      {
        "x": 761,
        "y": 8,
        "w": 154,
        "h": 189,
        "ax": 77,
        "ay": 189
      }
    ],
    "crouch": [
      {
        "x": 8,
        "y": 236,
        "w": 163,
        "h": 150,
        "ax": 81,
        "ay": 150
      },
      {
        "x": 259,
        "y": 236,
        "w": 153,
        "h": 147,
        "ax": 76,
        "ay": 147
      },
      {
        "x": 510,
        "y": 236,
        "w": 146,
        "h": 149,
        "ax": 72,
        "ay": 149
      },
      {
        "x": 761,
        "y": 236,
        "w": 153,
        "h": 148,
        "ax": 76,
        "ay": 148
      }
    ],
    "punchHigh": [
      {
        "x": 8,
        "y": 464,
        "w": 173,
        "h": 199,
        "ax": 86,
        "ay": 199
      },
      {
        "x": 259,
        "y": 464,
        "w": 214,
        "h": 184,
        "ax": 79,
        "ay": 184
      },
      {
        "x": 510,
        "y": 464,
        "w": 170,
        "h": 192,
        "ax": 85,
        "ay": 192
      },
      {
        "x": 761,
        "y": 464,
        "w": 154,
        "h": 192,
        "ax": 77,
        "ay": 192
      }
    ],
    "punchLow": [
      {
        "x": 8,
        "y": 692,
        "w": 174,
        "h": 159,
        "ax": 87,
        "ay": 159
      },
      {
        "x": 259,
        "y": 692,
        "w": 214,
        "h": 164,
        "ax": 90,
        "ay": 164
      },
      {
        "x": 510,
        "y": 692,
        "w": 186,
        "h": 166,
        "ax": 93,
        "ay": 166
      },
      {
        "x": 761,
        "y": 692,
        "w": 157,
        "h": 147,
        "ax": 78,
        "ay": 147
      }
    ],
    "kickHigh": [
      {
        "x": 8,
        "y": 920,
        "w": 113,
        "h": 203,
        "ax": 35,
        "ay": 203
      },
      {
        "x": 259,
        "y": 920,
        "w": 209,
        "h": 204,
        "ax": 53,
        "ay": 204
      },
      {
        "x": 510,
        "y": 920,
        "w": 116,
        "h": 212,
        "ax": 37,
        "ay": 212
      },
      {
        "x": 761,
        "y": 920,
        "w": 161,
        "h": 204,
        "ax": 80,
        "ay": 204
      }
    ],
    "kickLow": [
      {
        "x": 8,
        "y": 1148,
        "w": 165,
        "h": 201,
        "ax": 82,
        "ay": 201
      },
      {
        "x": 259,
        "y": 1148,
        "w": 153,
        "h": 193,
        "ax": 32,
        "ay": 193
      },
      {
        "x": 510,
        "y": 1148,
        "w": 197,
        "h": 197,
        "ax": 12,
        "ay": 197
      },
      {
        "x": 761,
        "y": 1148,
        "w": 174,
        "h": 156,
        "ax": 87,
        "ay": 156
      }
    ],
    "jump": [
      {
        "x": 8,
        "y": 1376,
        "w": 150,
        "h": 156,
        "ax": 75,
        "ay": 156
      },
      {
        "x": 259,
        "y": 1376,
        "w": 140,
        "h": 200,
        "ax": 56,
        "ay": 200
      },
      {
        "x": 510,
        "y": 1376,
        "w": 201,
        "h": 164,
        "ax": 19,
        "ay": 164
      },
      {
        "x": 761,
        "y": 1376,
        "w": 176,
        "h": 148,
        "ax": 88,
        "ay": 148
      }
    ],
    "ranged": [
      {
        "x": 8,
        "y": 1604,
        "w": 172,
        "h": 201,
        "ax": 86,
        "ay": 201
      },
      {
        "x": 259,
        "y": 1604,
        "w": 235,
        "h": 201,
        "ax": 85,
        "ay": 201
      },
      {
        "x": 510,
        "y": 1604,
        "w": 173,
        "h": 199,
        "ax": 83,
        "ay": 199
      },
      {
        "x": 761,
        "y": 1604,
        "w": 168,
        "h": 201,
        "ax": 84,
        "ay": 201
      }
    ]
  }
}